Clean up the task header a bit
