+++ src/api/transform/stream.ts
@@ -1,4 +1,5 @@
 export type ApiStream = AsyncGenerator<ApiStreamChunk>
+
 export type ApiStreamChunk = ApiStreamTextChunk | ApiStreamUsageChunk | ApiStreamReasoningChunk
 
 export interface ApiStreamTextChunk {
@@ -17,5 +18,6 @@ export interface ApiStreamUsageChunk {
 	outputTokens: number
 	cacheWriteTokens?: number
 	cacheReadTokens?: number
-	totalCost?: number // openrouter
+	reasoningTokens?: number
+	totalCost?: number
 }
