+++ webview-ui/src/components/chat/Mention.tsx
@@ -1,0 +1,33 @@
+import { mentionRegexGlobal } from "@roo/shared/context-mentions"
+
+import { vscode } from "../../utils/vscode"
+
+interface MentionProps {
+	text?: string
+	withShadow?: boolean
+}
+
+export const Mention = ({ text, withShadow = false }: MentionProps) => {
+	if (!text) {
+		return <>{text}</>
+	}
+
+	const parts = text.split(mentionRegexGlobal).map((part, index) => {
+		if (index % 2 === 0) {
+			// This is regular text.
+			return part
+		} else {
+			// This is a mention.
+			return (
+				<span
+					key={index}
+					className={`${withShadow ? "mention-context-highlight-with-shadow" : "mention-context-highlight"} cursor-pointer`}
+					onClick={() => vscode.postMessage({ type: "openMention", text: part })}>
+					@{part}
+				</span>
+			)
+		}
+	})
+
+	return <>{parts}</>
+}
